Bugfix/admin panel

+++ src/app/Traits/Fileable.php
@@ -3,92 +3,87 @@
 namespace App\Traits;
 
 use App\Models\Core\File as CoreFile;
-use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Arr;
 use Intervention\Image\Facades\Image;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Response;
 use Illuminate\Support\Facades\File;
-
+use App\Enums\StorageKey;
+use Illuminate\Http\UploadedFile;
 trait Fileable
 {
 
-
-    public $storage = [
-        "s3"      => "awsConfig",
-        "ftp"     => "ftpConfig"
-    ];
-
     /**
-     * store a file & image
+     * Store a file & image
      *
      * @param  $file
      * @param string $location
      * @param string $size
-     * @param string $removeFile
+     * @param CoreFile|null $removeFile
      * @param string $name
      * @return array
      */
-    public  function storeFile(mixed $file , string $location , string $size = null , mixed $removeFile = null ,string $name = null ): array 
-    {
+    public  function storeFile(UploadedFile $file , string $location , ? string $size = null ,  ? CoreFile $removeFile = null , ? string $name = null ): array{
+
 
         $name          = uniqid() . time() . '.' . $file->getClientOriginalExtension();
         $imagePath     = $location . '/' .$name ;
         $status        = true;
-        $message       = translate("File Uploaded Successfully");
+        $disk          = site_settings('storage');
     
-        if($removeFile){
-            $this->unlink($location ,$removeFile) ;    
-        }
-      
-        if(is_array($this->storage) && in_array(site_settings('storage'), array_keys($this->storage))){
-
-            $this->{$this->storage[site_settings('storage')]}();
-    
-            \Storage::disk(site_settings('storage'))->putFileAs(
-                $location ,
-                $file,
-                $name 
-            );
-
-        }
-        else{
-
-            if (!file_exists($location)) {
-                mkdir($location, 0755, true);
-            }
-
-            if(substr($file->getMimeType(), 0, 5) == 'image') {
-
-                $image = Image::make(file_get_contents($file));
-                if (isset($size)) {
-                    
-                    list($width, $height) = explode('x', strtolower($size));
-                    $image->resize($width, $height,function ($constraint) {
-                        $constraint->aspectRatio();
-                    });
-                   
+        //remove file if exists
+        if($removeFile) $this->unlink($location,$removeFile) ;    
+        switch ($disk) {
+            case StorageKey::LOCAL->value:
+                if (!file_exists($location))   mkdir($location, 0755, true);
+                switch (substr($file->getMimeType(), 0, 5)) {
+                    case 'image':
+                        $image = Image::make(file_get_contents($file));
+                        if (isset($size)) {
+                            list($width, $height) = explode('x', strtolower($size));
+                            $image->resize($width, $height, function ($constraint) :void{
+                                $constraint->aspectRatio();
+                            });
+                        }
+                        $image->save($imagePath);
+                        break;
+                
+                    default:
+                        $file->move($location, $name);
+                        break;
                 }
-                $image->save($imagePath);
-            }
-            else{
-                $file->move($location,$name);
-            }
+                break;
+        
+            default:
+                $this->{Arr::get(StorageKey::toArray(),strtoupper($disk))}();
+                \Storage::disk($disk)->putFileAs(
+                    $location ,
+                    $file,
+                    $name 
+                );
+                break;
         }
 
+    
         return [
             'status'     => $status,
-            'message'    => $message,
             'name'       => $name,
             'disk'       => site_settings('storage') ,
             "size"       => $this->formatSize( $file->getSize()),
             "extension"  => strtolower($file->getClientOriginalExtension())
         ];
-   
+
     }
 
 
-    function formatSize(string | int $bytes) :string {
+
+    /**
+     * Get file size
+     * 
+     * @param  int|string $bytes
+     * @return string
+     */
+    public function formatSize(string|int $bytes): string{
         $units = ['B', 'KB', 'MB', 'GB', 'TB'];
         for ($i = 0; (int) $bytes >= 1024 && $i < 4; $i++) {
             $bytes /= 1024;
@@ -98,96 +93,97 @@ trait Fileable
 
 
 
+
     /**
      * Unlink File Or Image 
      *
      * @param string $location
-     * @param mixed $file
-     * @return void
+     * @param CoreFile|null $file
+     * @return bool
      */
-    public function unlink(string $location ,mixed $file = null) :void{
+    public function unlink(string $location ,?CoreFile $file = null): bool{
 
 
         try {
-
-            if($file->disk ==  'local'){
-                if (file_exists($location . '/' . @$file->name) && is_file($location . '/' . @$file->name)) {
-                    @unlink($location . '/' . @$file->name);
-                }
-            }
-            else{
-
-                $this->{$this->storage[$file->disk]}();
-                
-                if(Storage::disk($file->disk)->exists($location . '/' . @$file->name)){
-                    Storage::disk($file->disk)->delete($location . '/' . @$file->name);
-                }
+            switch (@$file->disk) {
+                case StorageKey::LOCAL->value:
+                    if (file_exists($location . '/' . @$file->name) && is_file($location . '/' . @$file->name)) @unlink($location . '/' . @$file->name);
+                    break;
+            
+                default:
+                    $this->{Arr::get(StorageKey::toArray(), strtoupper($file->disk))}();
+                    if (Storage::disk($file->disk)->exists($location . '/' . @$file->name)) Storage::disk($file->disk)->delete($location . '/' . @$file->name);
+                    break;
             }
 
             @$file->delete();
-       
 
-           
-        } catch (\Throwable $th) {
-
+ 
+        } catch (\Exception $ex) {
+            return false;
         }
-       
 
-
+        return true;
     }
 
 
     /**
-     * aws configuration 
+     * Set aws configuration 
      *
      * @return void
      */
-    public function awsConfig() :void{
-        $aws_config = json_decode(site_settings('aws_s3'),true);
+    public function setAWSConfig() :void {
+      
+        $awsConfig = json_decode(site_settings('aws_s3'),true);
+
         config(
             [
-                'filesystems.disks.s3.key' => $aws_config['s3_key'],
-                'filesystems.disks.s3.secret' => $aws_config['s3_secret'],
-                'filesystems.disks.s3.region' => $aws_config['s3_region'],
-                'filesystems.disks.s3.bucket' => $aws_config['s3_bucket'],
+                'filesystems.disks.s3.key'                     => Arr::get($awsConfig ,'s3_key'),
+                'filesystems.disks.s3.secret'                  => Arr::get($awsConfig ,'s3_secret'),
+                'filesystems.disks.s3.region'                  => Arr::get($awsConfig ,'s3_region'),
+                'filesystems.disks.s3.bucket'                  => Arr::get($awsConfig ,'s3_bucket'),
                 'filesystems.disks.s3.use_path_style_endpoint' => false,
             ]
         );
     }
 
 
+
+
     /**
-     * ftp configuration 
+     * set ftp configuration 
      *
      * @return void
      */
-    public function ftpConfig() :void{
+    public function setFTPConfig() :void{
 
-        $ftp_config = json_decode(site_settings('ftp'),true);
+        $ftpConfig = json_decode(site_settings('ftp'),true);
         config(
             [
-                'filesystems.disks.ftp.host' => $ftp_config['host'],
-                'filesystems.disks.ftp.username' => $ftp_config['user_name'],
-                'filesystems.disks.ftp.password' => $ftp_config['password'],
-                'filesystems.disks.ftp.port' => (int) $ftp_config['port'],
-                'filesystems.disks.ftp.root' => $ftp_config['root']
+                'filesystems.disks.ftp.host'     => Arr::get($ftpConfig ,'host'),
+                'filesystems.disks.ftp.username' => Arr::get($ftpConfig ,'user_name') ,
+                'filesystems.disks.ftp.password' => Arr::get($ftpConfig ,'password'),
+                'filesystems.disks.ftp.port'     => (int) Arr::get($ftpConfig ,'port'),
+                'filesystems.disks.ftp.root'     =>  Arr::get($ftpConfig ,'root')
             ]
         );
 
     }
 
-
-    function getConfigValue($path, $config) {
-        $keys = explode(',', $path);
-    
+ /**
+     * get base URL
+     *
+     * @param string $path
+     * @param array $config
+     * @return array|null
+     */
+    function getBasePath(string $path, array $config): array|null {
+        
+        $keys    = explode(',', $path);
         $current = $config;
-    
         foreach ($keys as $key) {
-            if (isset($current[$key])) {
-                $current = $current[$key];
-            } else {
-                return null;
-            }
+            if(!Arr::exists($current, $key)) return null;
+            $current = $current[$key];
         }
     
         return $current;
@@ -195,50 +191,46 @@ trait Fileable
     
 
 
-
     /**
-     * get image url
+     * Get image URL
      *
-     * @param File $image
+     * @param CoreFile|null $image
      * @param string $path
      * @return string
      */
-    public function getImageUrl(mixed  $file, string $path ,bool $size , ?string $foreceSize  = null ) :string
-    {
+    public function getimageURL(CoreFile $file = null, string $path ,bool $size , ? string $foreceSize  = null): string{
         $config    = config("settings")['file_path'];
-        $basepath  = $this->getConfigValue($path, $config);
-        $image_url = asset('assets/images/default/default.jpg');
+        $basepath  = $this->getBasePath($path, $config);
+        $imageURL  = asset('assets/images/default/default.jpg');
 
         if($size){
             $default = $foreceSize?? "100x100";
-            $image_url = route('default.image',Arr::get($basepath,'size',$default));
+            $imageURL = route('default.image',Arr::get($basepath,'size',$default));
         }
-
         try {
 
-            if(isset($basepath['path'])){
-
+            if(Arr::exists($basepath, 'path')){
                 $image = $basepath['path']."/".@$file->name;
-                if(@$file->disk == "local"){
-                    if (file_exists($image) && is_file($image)) {
-                        $image_url =  asset($image);
-                    }
+                switch ($file->disk) {
+                    case StorageKey::LOCAL->value:
+                        if (file_exists($image) && is_file($image))  $imageURL =  asset($image);
+                        break;
+                    default:
+                        $this->{Arr::get(StorageKey::toArray(), strtoupper($file->disk))}();
+                        if(Storage::disk(@$file->disk)->exists($image))  $imageURL = \Storage::disk(@$file->disk)->url($image);
+                        break;
                 }
-                else{
-                    $this->{$this->storage[@$file->disk]}();
-                    if(Storage::disk(@$file->disk)->exists($image)){
-                        $image_url = \Storage::disk(@$file->disk)->url($image);
-                    }
-                }
-    
             }
          
         } catch (\Throwable $th) {
             
         }
-        return  $image_url;          
+        return  $imageURL;          
     }
-    
+
+
+
+ 
 
     /**
      * Check if file exists or not
@@ -246,8 +238,7 @@ trait Fileable
      * @param string $url
      * @return boolean
      */
-    public static function  check_file(string $url) :bool
-    {
+    public static function  check_file(string $url) :bool{
         $headers = get_headers($url);
         return (bool) preg_match('/\bContent-Type:\s+(?:image|audio|video)/i', implode("\n", $headers));
     }
@@ -255,39 +246,45 @@ trait Fileable
 
 
     /**
-     * download a file
+     * Download a file
      *
      * @param string $location
-     * @param mixed $file
+     * @param CoreFile|null $file
      * @return mixed
      */
-    public function downloadFile(string $location, mixed $file) :mixed
-    {
+    public function downloadFile(string $location, ?CoreFile $file = null): mixed{
+        
         $filePath = $location . '/' . $file->name;
-        $url = null;
+        $URL      = null;
 
         try {
 
-            if(is_array($this->storage) && in_array($file->disk, array_keys($this->storage))){
-                $this->{$this->storage[$file->disk]}();
-                $headers = [
-                    'Content-Disposition' => 'attachment; filename="'. $file->name .'"',
-                ];
+            switch ($file->disk) {
+                case StorageKey::LOCAL->value:
+                    $headers = [
+                        'Content-Type' => File::mimeType($filePath),
+                    ];
+                    $URL =  Response::download( $filePath,$file->name, $headers);
+                    break;
+            
+                default:
+                    $this->{Arr::get(StorageKey::toArray(), strtoupper($file->disk))}();
+            
+                    $headers = [
+                        'Content-Disposition' => 'attachment; filename="'. $file->name .'"',
+                    ];
 
-                $url =  Response::make(\Storage::disk($file->disk)->get($filePath),200, $headers);
+                    $URL =  Response::make(\Storage::disk($file->disk)->get($filePath),200, $headers);
+                    break;
             }
 
-            else{
-                $headers = [
-                    'Content-Type' => File::mimeType($filePath),
-                ];
-                $url =  Response::download( $filePath,$file->name, $headers);
-            }
+
         } catch (\Throwable $th) {
 
         }
 
-        return $url;
+        return $URL;
+
     }
